black-friday/answers/requirements.txt
appnope==0.1.0
astroid==2.3.2
astropy==3.2.3
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cloudpickle==1.2.2
confuse==1.0.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
google-api-core==1.14.3
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.21.0
google-cloud-core==1.0.3
google-cloud-storage==1.21.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
htmlmin==0.1.12
hyperopt==0.2.2
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
json5==0.8.5
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.1
jupyterlab-server==1.0.6
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
llvmlite==0.30.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.2
notebook==6.0.2
numba==0.46.0
numpy==1.17.3
oauthlib==3.1.0
packaging==19.2
pandas==0.25.3
pandas-gbq==0.11.0
pandas-profiling==2.3.0
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
phik==0.9.8
pickleshare==0.7.5
pluggy==0.13.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pydata-google-auth==0.1.3
Pygments==2.4.2
pylint==2.4.3
pyparsing==2.4.3
pyrsistent==0.15.5
pytest==5.2.2
pytest-pylint==0.14.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
terminado==0.8.2
testpath==0.4.4
tornado==6.0.3
tqdm==4.37.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.2
xgboost==0.90
zipp==0.6.0